requirements_base.txt
Django>1.7,<1.8
times
pytz
djangorestframework
django-filter
django-rq
